pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jobsgrep"
version = "1.0.0"
description = "Self-hosted job search web app with legal-first API architecture"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "openpyxl>=3.1.2",
    "groq>=0.9.0",
    "pycryptodome>=3.20.0",
    "cryptography>=42.0.0",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.1.0",
    "aiofiles>=23.2.1",
    "PyMuPDF>=1.24.0",
]

[project.optional-dependencies]
scraping = [
    "python-jobspy>=1.1.80",
]
redis = [
    "redis>=5.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[project.scripts]
jobsgrep = "jobsgrep.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["jobsgrep*"]
